--- src/app/base.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-    <meta charset="UTF-8">
-    <title>Person</title>
-</head>
-<body>
-{% block content %}{% endblock %}
-</body>
-</html>
