+++ src/app/urls.py
@@ -1,22 +1,11 @@
 # -*- coding:utf-8 -*-
 
 from django.conf.urls import include, url
-#from app.views import contact
-from app.views import ContactView
-from app.test_views import TestRequest
-#from app.views import PersonCreateView
-#from app.views import hello_template
-#from app.views import hello_forms2
-#from app.views import ContactView 
+from app.views.views import ContactView
+from app.views.test_views import TestRequest
 from django.contrib import admin
 
 urlpatterns = [
-#    url(r'^mail/', contact),
-#	url(r'create/$', PersonCreateView.as_view()),
-#	url(r'template$', hello_template, name='hello_template'),
-#    url(r'^form_samples/$', hello_forms2, name='hello_forms2'),
-#	url(r'^contact/$', ContactView.as_view(), name='ContactView'),
-#	url(r'^contact2/$', ContactView.as_view(), name='ContactView'),
 	url(r'^thanks/$', ContactView.get_name),
 	url(r'^contact/$', ContactView.get_name),
 	url(r'^test/$', TestRequest.test_request),
